refactor(wallet): Remove deprecated service worker implementations and update related imports
