+++ src/identity/index.ts
@@ -10,5 +10,3 @@ export interface Identity {
 }
 
 export * from "./singleKey";
-export * from "./serviceWorker";
-export * from "./proxyIdentity";
