--- src/wallet/serviceWorker/wallet.ts
@@ -1,377 +1,0 @@
-import {
-    IWallet,
-    WalletBalance,
-    SendBitcoinParams,
-    SettleParams,
-    ArkTransaction,
-    ExtendedCoin,
-    ExtendedVirtualCoin,
-    GetVtxosFilter,
-} from "..";
-import { Request } from "./request";
-import { Response } from "./response";
-import { SettlementEvent } from "../../providers/ark";
-import { hex } from "@scure/base";
-import { Identity } from "../../identity";
-import { StorageAdapter } from "../../storage";
-import { IndexedDBStorageAdapter } from "../../storage/indexedDB";
-import { WalletRepository } from "../../repositories/walletRepository";
-import { WalletRepositoryImpl } from "../../repositories/walletRepository";
-import { ContractRepository } from "../../repositories/contractRepository";
-import { ContractRepositoryImpl } from "../../repositories/contractRepository";
-
-class UnexpectedResponseError extends Error {
-    constructor(response: Response.Base) {
-        super(
-            `Unexpected response type. Got: ${JSON.stringify(response, null, 2)}`
-        );
-        this.name = "UnexpectedResponseError";
-    }
-}
-
-/**
- * Service Worker-based wallet implementation for browser environments.
- *
- * @deprecated Use Wallet instead. This class will be removed in a future version.
- * The new Wallet class automatically detects execution context and handles service worker
- * initialization transparently, eliminating the need for manual service worker setup.
- *
- * Migration guide:
- * ```typescript
- * // Old approach - manual service worker setup
- * const serviceWorker = await setupServiceWorker("/service-worker.js");
- * const identity = new ServiceWorkerIdentity(serviceWorker);
- * const wallet = await ServiceWorkerWallet.create({
- *   serviceWorker,
- *   identity,
- *   arkServerUrl: 'https://ark.example.com'
- * });
- *
- * // New approach - automatic context detection
- * import { Wallet, SingleKey } from '@arkade-os/sdk';
- * const wallet = await Wallet.create({
- *   identity: SingleKey.fromHex('your_key'), // or any Identity implementation
- *   arkServerUrl: 'https://ark.example.com'
- * });
- * ```
- *
- * The new Wallet class automatically:
- * - Detects if running in a service worker context
- * - Sets up service worker communication when needed
- * - Falls back to direct execution in Node.js or when service workers aren't available
- * - Manages persistent identity storage in service worker environments
- *
- * This legacy wallet uses a service worker as a backend to handle wallet logic,
- * providing secure key storage and transaction signing in web applications.
- * The service worker runs in a separate thread and can persist data between
- * browser sessions.
- *
- * @example
- * ```typescript
- * // Legacy approach (deprecated)
- * const serviceWorker = await setupServiceWorker("/service-worker.js");
- * const identity = new ServiceWorkerIdentity(serviceWorker);
- * const wallet = await ServiceWorkerWallet.create({
- *   serviceWorker,
- *   identity,
- *   arkServerUrl: 'https://ark.example.com'
- * });
- *
- * // Use like any other wallet
- * const address = await wallet.getAddress();
- * const balance = await wallet.getBalance();
- * ```
- */
-export interface ServiceWorkerWalletCreateOptions {
-    serviceWorker: ServiceWorker;
-    identity: Identity;
-    arkServerUrl: string;
-    esploraUrl?: string;
-    arkServerPublicKey?: string;
-    storage?: StorageAdapter;
-    privateKey?: string; // Optional private key for initial identity setup
-}
-
-export class ServiceWorkerWallet implements IWallet {
-    public readonly walletRepository: WalletRepository;
-    public readonly contractRepository: ContractRepository;
-    public readonly identity: Identity;
-
-    private constructor(
-        public readonly serviceWorker: ServiceWorker,
-        identity: Identity,
-        walletRepository: WalletRepository,
-        contractRepository: ContractRepository
-    ) {
-        this.identity = identity;
-        this.walletRepository = walletRepository;
-        this.contractRepository = contractRepository;
-    }
-
-    static async create(
-        options: ServiceWorkerWalletCreateOptions
-    ): Promise<ServiceWorkerWallet> {
-        // Default to IndexedDB for service worker context
-        const storage =
-            options.storage || new IndexedDBStorageAdapter("wallet-db");
-
-        // Create repositories
-        const walletRepo = new WalletRepositoryImpl(storage);
-        const contractRepo = new ContractRepositoryImpl(storage);
-
-        // Initialize the service worker with the config
-        // Note: Service worker manages its own persistent identity stored in IndexedDB
-        // When privateKey is null, the service worker will load existing identity or generate a new one
-        const initMessage: Request.InitWallet = {
-            type: "INIT_WALLET",
-            id: getRandomId(),
-            privateKey: options.privateKey,
-            arkServerUrl: options.arkServerUrl,
-            arkServerPublicKey: options.arkServerPublicKey,
-        };
-
-        const wallet = new ServiceWorkerWallet(
-            options.serviceWorker,
-            options.identity,
-            walletRepo,
-            contractRepo
-        );
-
-        // Initialize the service worker
-        await wallet.sendMessage(initMessage);
-
-        return wallet;
-    }
-
-    // send a message and wait for a response
-    private async sendMessage<T extends Request.Base>(
-        message: T
-    ): Promise<Response.Base> {
-        return new Promise((resolve, reject) => {
-            const messageHandler = (event: MessageEvent) => {
-                const response = event.data as Response.Base;
-                if (response.id === "") {
-                    reject(new Error("Invalid response id"));
-                    return;
-                }
-                if (response.id !== message.id) {
-                    return;
-                }
-                navigator.serviceWorker.removeEventListener(
-                    "message",
-                    messageHandler
-                );
-
-                if (!response.success) {
-                    reject(new Error((response as Response.Error).message));
-                } else {
-                    resolve(response);
-                }
-            };
-
-            navigator.serviceWorker.addEventListener("message", messageHandler);
-            this.serviceWorker.postMessage(message);
-        });
-    }
-
-    async clear() {
-        const message: Request.Clear = {
-            type: "CLEAR",
-            id: getRandomId(),
-        };
-        await this.sendMessage(message);
-    }
-
-    async getAddress(): Promise<string> {
-        const message: Request.GetAddress = {
-            type: "GET_ADDRESS",
-            id: getRandomId(),
-        };
-
-        try {
-            const response = await this.sendMessage(message);
-            if (Response.isAddress(response)) {
-                return response.address;
-            }
-            throw new UnexpectedResponseError(response);
-        } catch (error) {
-            throw new Error(`Failed to get address: ${error}`);
-        }
-    }
-
-    async getBoardingAddress(): Promise<string> {
-        const message: Request.GetBoardingAddress = {
-            type: "GET_BOARDING_ADDRESS",
-            id: getRandomId(),
-        };
-
-        try {
-            const response = await this.sendMessage(message);
-            if (Response.isBoardingAddress(response)) {
-                return response.address;
-            }
-            throw new UnexpectedResponseError(response);
-        } catch (error) {
-            throw new Error(`Failed to get boarding address: ${error}`);
-        }
-    }
-
-    async getBalance(): Promise<WalletBalance> {
-        const message: Request.GetBalance = {
-            type: "GET_BALANCE",
-            id: getRandomId(),
-        };
-
-        try {
-            const response = await this.sendMessage(message);
-            if (Response.isBalance(response)) {
-                return response.balance;
-            }
-            throw new UnexpectedResponseError(response);
-        } catch (error) {
-            throw new Error(`Failed to get balance: ${error}`);
-        }
-    }
-
-    async getBoardingUtxos(): Promise<ExtendedCoin[]> {
-        const message: Request.GetBoardingUtxos = {
-            type: "GET_BOARDING_UTXOS",
-            id: getRandomId(),
-        };
-
-        try {
-            const response = await this.sendMessage(message);
-            if (Response.isBoardingUtxos(response)) {
-                return response.boardingUtxos;
-            }
-            throw new UnexpectedResponseError(response);
-        } catch (error) {
-            throw new Error(`Failed to get boarding UTXOs: ${error}`);
-        }
-    }
-
-    async getStatus(): Promise<Response.WalletStatus["status"]> {
-        const message: Request.GetStatus = {
-            type: "GET_STATUS",
-            id: getRandomId(),
-        };
-        const response = await this.sendMessage(message);
-        if (Response.isWalletStatus(response)) {
-            return response.status;
-        }
-        throw new UnexpectedResponseError(response);
-    }
-
-    async getTransactionHistory(): Promise<ArkTransaction[]> {
-        const message: Request.GetTransactionHistory = {
-            type: "GET_TRANSACTION_HISTORY",
-            id: getRandomId(),
-        };
-
-        try {
-            const response = await this.sendMessage(message);
-            if (Response.isTransactionHistory(response)) {
-                return response.transactions;
-            }
-            throw new UnexpectedResponseError(response);
-        } catch (error) {
-            throw new Error(`Failed to get transaction history: ${error}`);
-        }
-    }
-
-    async getVtxos(filter?: GetVtxosFilter): Promise<ExtendedVirtualCoin[]> {
-        const message: Request.GetVtxos = {
-            type: "GET_VTXOS",
-            id: getRandomId(),
-            filter,
-        };
-
-        try {
-            const response = await this.sendMessage(message);
-            if (Response.isVtxos(response)) {
-                return response.vtxos;
-            }
-            throw new UnexpectedResponseError(response);
-        } catch (error) {
-            throw new Error(`Failed to get vtxos: ${error}`);
-        }
-    }
-
-    async sendBitcoin(params: SendBitcoinParams): Promise<string> {
-        const message: Request.SendBitcoin = {
-            type: "SEND_BITCOIN",
-            params,
-            id: getRandomId(),
-        };
-
-        try {
-            const response = await this.sendMessage(message);
-            if (Response.isSendBitcoinSuccess(response)) {
-                return response.txid;
-            }
-            throw new UnexpectedResponseError(response);
-        } catch (error) {
-            throw new Error(`Failed to send bitcoin: ${error}`);
-        }
-    }
-
-    async settle(
-        params?: SettleParams,
-        callback?: (event: SettlementEvent) => void
-    ): Promise<string> {
-        const message: Request.Settle = {
-            type: "SETTLE",
-            params,
-            id: getRandomId(),
-        };
-
-        try {
-            return new Promise((resolve, reject) => {
-                const messageHandler = (event: MessageEvent) => {
-                    const response = event.data as Response.Base;
-
-                    if (!response.success) {
-                        navigator.serviceWorker.removeEventListener(
-                            "message",
-                            messageHandler
-                        );
-                        reject(new Error((response as Response.Error).message));
-                        return;
-                    }
-
-                    switch (response.type) {
-                        case "SETTLE_EVENT":
-                            if (callback) {
-                                callback(
-                                    (response as Response.SettleEvent).event
-                                );
-                            }
-                            break;
-                        case "SETTLE_SUCCESS":
-                            navigator.serviceWorker.removeEventListener(
-                                "message",
-                                messageHandler
-                            );
-                            resolve((response as Response.SettleSuccess).txid);
-                            break;
-                        default:
-                            break;
-                    }
-                };
-
-                navigator.serviceWorker.addEventListener(
-                    "message",
-                    messageHandler
-                );
-                this.serviceWorker.postMessage(message);
-            });
-        } catch (error) {
-            throw new Error(`Settlement failed: ${error}`);
-        }
-    }
-}
-
-function getRandomId(): string {
-    const randomValue = crypto.getRandomValues(new Uint8Array(16));
-    return hex.encode(randomValue);
-}
